Add files via upload
Hey, this game is based on a monkey who is very hungry, he wants some food to eat can you help him.
How to play?? When you will click on this link your game will be started automatically. When you press the spacebar the monkey will jump.
The aim of this game is to get monkey some food and win points as much as you can.